+++ src/Manager/Environment/effects/removeDir.ts
@@ -6,16 +6,11 @@
 
 import fse from 'fs-extra';
 import path from 'path';
-import { ErrorDialogActions, usageData } from 'pc-nrfconnect-shared';
 
-import { Dispatch } from '../../../state';
-
-export const removeDir = async (dispatch: Dispatch, srcDir: string) => {
-    let renameOfDirSuccessful = false;
+export const removeDir = async (srcDir: string) => {
+    const toBeDeletedDir = path.resolve(srcDir, '..', 'toBeDeleted');
     try {
-        const toBeDeletedDir = path.resolve(srcDir, '..', 'toBeDeleted');
-        await fse.move(srcDir, toBeDeletedDir, { overwrite: true });
-        renameOfDirSuccessful = true;
+        await fse.rename(srcDir, toBeDeletedDir);
         await fse.remove(toBeDeletedDir);
     } catch (error) {
         const [, , message] = `${error}`.split(/[:,] /);
@@ -23,8 +18,7 @@ export const removeDir = async (dispatch: Dispatch, srcDir: string) => {
             `Failed to remove ${srcDir}, ${message}. ` +
             'Please close any application or window that might keep this ' +
             'environment locked, then try to remove it again.';
-        dispatch(ErrorDialogActions.showDialog(errorMsg));
-        usageData.sendErrorReport(errorMsg);
+
+        throw new Error(errorMsg);
     }
-    return renameOfDirSuccessful;
 };